+++ notUsed/Weapon6.java
@@ -19,11 +19,11 @@ public class Weapon6 extends NormalWeapon{
             if (input==null) return;
             Vec2d pos=owner.getBulletPosition();
             Vec2d velocity=input.limit(speed*3);
-            float size=this.size*0.7;
-            float health=this.health*0.5;
-            float damage=this.damage*0.5;
+            double size=this.size*0.7;
+            double health=this.health*0.5;
+            double damage=this.damage*0.5;
             for(int i=0;i<30;i++){
-                float r=Util.random(0.8,1.5);
+                double r=Util.random(0.8,1.5);
                 shootBullet(pos,velocity.multiply(Util.random(1,3)).add(Util.randomVec().limit(0.13)),size*r,health*r,damage*r);
             }
